LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A6.4_T1-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Unexpected token ';'. Expected ')' to end a for-loop header.
S7.9_A6.4_T1

PASS Expected parsing failure

TEST COMPLETE
